+++ src/components/ButtonGroup.vue
@@ -7,46 +7,72 @@
 </template>
 
 <script lang="ts" setup>
-  import { ref, type PropType, computed } from 'vue'
+import { ref, type PropType, computed, defineEmits, watch } from 'vue'
 
-  type Button = {
-    text: string
-    value: any
+type Button = {
+  text: string
+  value: any
+}
+
+const props = defineProps({
+  buttons: {
+    type: Array as PropType<Button[]>,
+    required: true
+  },
+  allowMultipleSelection: {
+    type: Boolean,
+    default: false
+  },
+  initialValue: {
+    type: String,
+    default: null
   }
+})
 
-  const props = defineProps({
-    buttons: {
-      type: Array as PropType<Button[]>,
-      required: true
-    },
-    allowMultipleSelection: {
-      type: Boolean,
-      default: false
-    }
-  })
-  const selectedButtons = ref<number[]>([])
+const selectedButtons = ref<number[]>([])
 
-  function selectButton(index: number) {
-    if (props.allowMultipleSelection) {
-      const isSelected = selectedButtons.value.includes(index)
+function selectButton(index: number) {
+  if (props.allowMultipleSelection) {
+    const isSelected = selectedButtons.value.includes(index)
 
-      if (isSelected) {
-        selectedButtons.value = selectedButtons.value.filter(i => i !== index)
-      } else {
-        selectedButtons.value.push(index)
-      }
+    if (isSelected) {
+      selectedButtons.value = selectedButtons.value.filter(i => i !== index)
     } else {
-      selectedButtons.value = [index]
+      selectedButtons.value.push(index)
     }
-    emit('change', selectedValues.value)
+  } else {
+    selectedButtons.value = [index]
   }
 
-  const selectedValues = computed(() => {
-    return selectedButtons.value.map(index => props.buttons[index].value)
-  })
+  emit('change', selectedValues.value)
+}
 
-  const emit = defineEmits(['change'])
+const selectedValues = computed(() => {
+  return selectedButtons.value.map(index => props.buttons[index].value)
+})
 
+const emit = defineEmits(['change'])
+
+if (props.initialValue) {
+  const initialValueIndex = props.buttons.findIndex(button => button.value === props.initialValue)
+  if (initialValueIndex >= 0) {
+    selectedButtons.value.push(initialValueIndex)
+  }
+}
+
+watch(() => props.initialValue, (newValue, oldValue) => {
+  if(newValue == '') {
+    return
+  }
+  const selectedIndex = props.buttons.findIndex(button => button.value == newValue)
+  const selectedValue = (selectedIndex >= 0 ? props.buttons[selectedIndex] : '')
+  if (newValue !== oldValue || newValue != selectedValue) {
+    const newSelectedIndex = props.buttons.findIndex(button => button.value === newValue)
+    if (newSelectedIndex >= 0) {
+      selectedButtons.value = [newSelectedIndex]
+    }
+  }
+})
 </script>
 
 <style scoped>