+++ src/views/catprez/CatPrezHomeView.vue
@@ -12,6 +12,11 @@ const links = [
 
 <template>
     <FlavourHome flavour="CatPrez" :links="links">
+        <p>
+            A general data catalog conforming to <a href="https://www.w3.org/TR/vocab-dcat-2/">DCAT</a>, delivered 
+            as a read-only web system in both human and machine-readable views.
+            The views are known as profiles and a list of them are available <RouterLink to="/s/profiles">here</RouterLink>.
+        </p>
         <p>Welcome to CatPrez - a read-only web delivery system for Data Catalog Vocabulary (DCAT) information.</p>
     </FlavourHome>
 </template>